Migration from visual studio to CLion, small bugfixes
