--- SweepMethod.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31205.134
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SweepMethod", "SweepMethod\SweepMethod.vcxproj", "{C5366DE9-6169-4EDF-AB35-3744585F6E9E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C5366DE9-6169-4EDF-AB35-3744585F6E9E}.Debug|x64.ActiveCfg = Debug|x64
-		{C5366DE9-6169-4EDF-AB35-3744585F6E9E}.Debug|x64.Build.0 = Debug|x64
-		{C5366DE9-6169-4EDF-AB35-3744585F6E9E}.Debug|x86.ActiveCfg = Debug|Win32
-		{C5366DE9-6169-4EDF-AB35-3744585F6E9E}.Debug|x86.Build.0 = Debug|Win32
-		{C5366DE9-6169-4EDF-AB35-3744585F6E9E}.Release|x64.ActiveCfg = Release|x64
-		{C5366DE9-6169-4EDF-AB35-3744585F6E9E}.Release|x64.Build.0 = Release|x64
-		{C5366DE9-6169-4EDF-AB35-3744585F6E9E}.Release|x86.ActiveCfg = Release|Win32
-		{C5366DE9-6169-4EDF-AB35-3744585F6E9E}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {3A0EB5D3-88C9-4E44-82D4-596DE5B60AB3}
-	EndGlobalSection
-EndGlobal
