--- SweepMethod/SweepMethod.vcxproj.filters
@@ -1,90 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Source Files">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;c++;cppm;ixx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Header Files">
-      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-      <Extensions>h;hh;hpp;hxx;h++;hm;inl;inc;ipp;xsd</Extensions>
-    </Filter>
-    <Filter Include="Resource Files">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-    <Filter Include="Header Files\interfaces">
-      <UniqueIdentifier>{795f52e3-b2e8-4267-bc77-a39ca2300d9d}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Header Files\parallel">
-      <UniqueIdentifier>{fbe37946-a9d4-4f7d-9ee9-65a99bad734a}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Header Files\serial">
-      <UniqueIdentifier>{f96f0d7e-d0e9-431d-81c7-2eff3bcd8c93}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Header Files\tools">
-      <UniqueIdentifier>{544477cc-8a6d-49a9-bbb5-3d72b5fb5544}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Header Files\tests">
-      <UniqueIdentifier>{9f451d45-486b-4343-8990-bc6d55db3ee5}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Source Files\tools">
-      <UniqueIdentifier>{eb14986f-33c2-4083-8c8b-770bd6fb6507}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Source Files\parallel">
-      <UniqueIdentifier>{b86e6d06-adf9-4898-91a6-4e27c12c0870}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Source Files\serial">
-      <UniqueIdentifier>{59ef16cd-b905-4525-bd68-6f8b79ed3cc7}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="main.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="Instrumental.cpp">
-      <Filter>Source Files\tools</Filter>
-    </ClCompile>
-    <ClCompile Include="ParallelInstrumental.cpp">
-      <Filter>Source Files\parallel</Filter>
-    </ClCompile>
-    <ClCompile Include="ParallelSweepMethod.cpp">
-      <Filter>Source Files\parallel</Filter>
-    </ClCompile>
-    <ClCompile Include="SerialSweepMethod.cpp">
-      <Filter>Source Files\serial</Filter>
-    </ClCompile>
-    <ClCompile Include="SerialInstrumental.cpp">
-      <Filter>Source Files\serial</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="AbstractSweepMethod.h">
-      <Filter>Header Files\interfaces</Filter>
-    </ClInclude>
-    <ClInclude Include="ParallelInstrumental.h">
-      <Filter>Header Files\parallel</Filter>
-    </ClInclude>
-    <ClInclude Include="ParallelSweepMethod.h">
-      <Filter>Header Files\parallel</Filter>
-    </ClInclude>
-    <ClInclude Include="SerialSweepMethod.h">
-      <Filter>Header Files\serial</Filter>
-    </ClInclude>
-    <ClInclude Include="Instrumental.h">
-      <Filter>Header Files\tools</Filter>
-    </ClInclude>
-    <ClInclude Include="Profiler.h">
-      <Filter>Header Files\tools</Filter>
-    </ClInclude>
-    <ClInclude Include="TestRunner.h">
-      <Filter>Header Files\tools</Filter>
-    </ClInclude>
-    <ClInclude Include="UnitTests.h">
-      <Filter>Header Files\tests</Filter>
-    </ClInclude>
-    <ClInclude Include="SerialInstrumental.h">
-      <Filter>Header Files\serial</Filter>
-    </ClInclude>
-  </ItemGroup>
-</Project>
