--- SweepMethod/main.cpp
@@ -1,9 +1,0 @@
-#include "UnitTests.h"
-
-
-int main(int argc, char** argv) {
-	UnitTests tests;
-	tests.execute();
-	
-	return 0;
-}
