+++ SweepMethod/main/Application.cpp
@@ -1,0 +1,9 @@
+#include "test/UnitTests.h"
+
+
+int main(int argc, char** argv) {
+	UnitTests tests;
+	tests.execute();
+	
+	return 0;
+}
