+++ SweepMethod/main/impl/parallel/ParallelSweepMethod.cpp
@@ -1,9 +1,9 @@
-#include "ParallelSweepMethod.h"
+#include "main/interfaces/parallel/ParallelSweepMethod.h"
 
 
 void ParallelSweepMethod::transformMatrA() {
 	for (int iter = 0; iter < THREAD_NUM; iter++) {
-		// сверху-вниз
+		// top-down
 		for (int i = iter * BLOCK_SIZE; i < (iter + 1) * BLOCK_SIZE - 1; i++) {
 			double k = A[i + 1][i] / A[i][i];
 			for (int j = 0; j < i + 1; j++) {
@@ -11,7 +11,7 @@ void ParallelSweepMethod::transformMatrA() {
 			}
 		}
 
-		// снизу-вверх
+		// bottom-up
 		for (int i = (iter + 1) * BLOCK_SIZE - 2; i >= iter * BLOCK_SIZE; i--) {
 			double k = A[i][i + 1] / A[i + 1][i + 1];
 			for (int j = i; j < N - 1; j++) {
@@ -25,28 +25,28 @@ std::pair<matr, vec> ParallelSweepMethod::collectInterferElem() {
 	matr R(CLASSIC_SIZE, vec(CLASSIC_SIZE, 0.));
 	vec  Y1(CLASSIC_SIZE);
 
-	// Находим вектор неизвестных
+	// finding vector of unknowns
 	for (int i = BLOCK_SIZE, k = 1; i < N; i += BLOCK_SIZE, k += 2) {
 		Y1[k - 1] = b[i - 1];
 		Y1[k] = b[i];
 	}
 
-	// 1. предобработка (крайние части)
-	// -> верхняя
+    // 1. preprocessing (extreme parts)
+    // -> upper
 	int Nb = BLOCK_SIZE - 1;
 	R[0][0] = A[Nb][Nb];
 	R[0][1] = A[Nb][BLOCK_SIZE];
 	R[1][0] = A[BLOCK_SIZE][Nb];
 	R[2][1] = A[2 * BLOCK_SIZE - 1][Nb];
 
-	// -> нижняя
+	// -> lower
 	Nb = N - BLOCK_SIZE;
 	R[CLASSIC_SIZE - 1][CLASSIC_SIZE - 1] = A[Nb][Nb];
 	R[CLASSIC_SIZE - 1][CLASSIC_SIZE - 2] = A[Nb][Nb - 1];
 	R[CLASSIC_SIZE - 2][CLASSIC_SIZE - 1] = A[Nb - 1][Nb];
 	R[CLASSIC_SIZE - 3][CLASSIC_SIZE - 2] = A[N - BLOCK_SIZE * 2][Nb];
 
-	// 2. постобработка (внутренняя часть)
+	// 2. post-processing (internal part)
 	for (int i = BLOCK_SIZE, k = 1; i < N - BLOCK_SIZE; i += BLOCK_SIZE, k += 2) {
 		for (int j = BLOCK_SIZE, l = 1; j < N - BLOCK_SIZE; j += BLOCK_SIZE, l += 2) {
 			double a = A[i][j];
@@ -72,7 +72,7 @@ std::pair<matr, vec> ParallelSweepMethod::collectInterferElem() {
 	printMatr(R, "R1");
 	printVec(Y1, "Y1");
 
-	// упорядочиваем коэф-ты
+	// ordering the coefficient
 	for (int i = 0; i < CLASSIC_SIZE; i += 2) {
 		std::swap(R[i][i], R[i][i + 1]);
 		std::swap(R[i + 1][i], R[i + 1][i + 1]);
@@ -117,7 +117,6 @@ vec ParallelSweepMethod::ClassicSweepMethod(const matr& R, const vec& Y1) {
 	return ssm.run();
 }
 
-// Получаем Y2 из векторов b, Y1
 vec ParallelSweepMethod::getY2(vec& Y1) {
 	vec Y2;
 
@@ -138,28 +137,28 @@ vec ParallelSweepMethod::getY2(vec& Y1) {
 vec ParallelSweepMethod::getX2(vec& Y2) {
 	vec X2(N - CLASSIC_SIZE);
 
-	// 1. предобработка (крайние части)
-	// -> верхняя + нижняя
+    // 1. preprocessing (extreme parts)
+    // -> upper + lower
 	int last = N - BLOCK_SIZE - 1;
 	for (int i = 0; i < BLOCK_SIZE - 1; i++) {
 		int j = N - i - 1;
 
-		// находим коэф-ты
+		// finding coefficients
 		Y2[i] -= A[i][BLOCK_SIZE];
 		Y2[j - CLASSIC_SIZE] -= A[j][last];
 
-		// находим вектор неизвестных
+		// finding vector of unknowns
 		X2[i] = Y2[i] / A[i][i];
 		X2[j - CLASSIC_SIZE] = Y2[j - CLASSIC_SIZE] / A[j][j];
 	}
 
-	// 2. постобработка (внутреняя часть)
+	// 2. post-processing (internal part)
 	for (int k = BLOCK_SIZE + 1, l = 1; k < N - BLOCK_SIZE - 1; k += BLOCK_SIZE, l++) {
 		for (int i = 0; i < BLOCK_SIZE - 2; i++) {
-			// находим коэф-ты
+			// finding coefficients
 			Y2[i + k - 2 * l] -= (A[i + k][k - 2] + A[i + k][k + BLOCK_SIZE - 1]);
 
-			// находим вектор неизвестных
+			// finding vector of unknowns
 			X2[i + k - 2 * l] = Y2[i + k - 2 * l] / A[i + k][i + k];
 		}
 	}
@@ -170,8 +169,8 @@ vec ParallelSweepMethod::getX2(vec& Y2) {
 vec ParallelSweepMethod::result(const vec& X1, const vec& X2) {
 	vec X(N);
 
-	// 1. предобработка (крайние части)
-	// -> верхняя + нижняя
+    // 1. preprocessing (extreme parts)
+    // -> upper + lower
 	for (int i = 0; i < BLOCK_SIZE - 1; i++) {
 		int j = N - i - 1;
 
@@ -183,7 +182,7 @@ vec ParallelSweepMethod::result(const vec& X1, const vec& X2) {
 		j = BLOCK_SIZE - 1,
 		k = BLOCK_SIZE - 1;
 
-	// 2. постобработка (внутреняя часть)
+    // 2. post-processing (internal part)
 	while (k < N - BLOCK_SIZE) {
 		for (int jiter = 0; jiter < 2; jiter++) {
 			X[k++] = X1[i++];
@@ -197,35 +196,34 @@ vec ParallelSweepMethod::result(const vec& X1, const vec& X2) {
 	return X;
 }
 
-// Вызов всех методов
 vec ParallelSweepMethod::run() {
 	printMatr(A, "A1");
 	printVec(b, "b");
 
-	// 1. Приведение матрицы к столбцовым видам
+    // 1. Reduction of the matrix to columnar views
 	transformMatrA();
 
 	printMatr(A, "A2");
 
-	// 2. Формируем матрицу R для последовательной прогонки вместе с вектором неизвестных Y1
+    // 2. From the matrix R for a serial sweep method along with the vector of unknowns Y1
 	matr R; vec Y1;
 	std::tie(R, Y1) = collectInterferElem();
 
 	printMatr(R, "R2");
 	printVec(Y1, "Y1");
 
-	// 3. Решаем СЛАУ (R*X1 = Y1) классической прогонкой
+	// 3. Solve SLAU (R * X1 = Y1) by the classical sweep method
 	vec X1 = ClassicSweepMethod(R, Y1);
 	printVec(X1, "X1");
 
-	// 4. Получаем вектор неизвестных Y2
+	// 4. Get the vector of unknowns Y2
 	vec Y2 = getY2(Y1);
 	printVec(Y2, "Y2");
 
-	// 5. Находим вектор неизвестных X2
+	// 5. Find the vector of unknowns X2
 	vec X2 = getX2(Y2);
 	printVec(X2, "X2");
 
-	// 6. Группируем вектора неизвестных X1, X2 в X
+	// 6. Group vectors of unknowns X1, X2 into X
 	return result(X1, X2);
 }
