+++ SweepMethod/main/interfaces/AbstractSweepMethod.h
@@ -6,5 +6,6 @@
 
 class AbstractSweepMethod {
 public:
+    // Calling sweep method
 	virtual std::vector<double> run() = 0;
 };
